Add files via upload
Just commiting this as a kind of example case. It's not done in an object oriented format, but i figured simply getting something with some progress would be useful.
Check it out. It's just the stage, displaying of the bumper and ball, and it has bumper moving and collision with the side walls.
Also, it's all done relatively, so changing the resolution values, or window size, should scale the elements dynamically.