Added in a little OOP and wall collision
Though, i think we need a better framework still. I went ahead and added in the wall collision and took some of the object details like position, velocity, and color, into the bb_objs.py file, which i import into the game.py file. 
Also cleaned up the paddle movement so that releasing the keys doesn't set the paddle velocity to 0, but instead reduces it. Basically it's not as clunky when rapidly pressing keys.