Home screen is set up, res folder added
The home screen is ready to link to any of our other files, "views", "activities", "modules", whatever they are called in python. Each of the buttons is set up to a function when clicked, exit is the only one currently that does anything though. I also created a res folder that has a couple of things in it(a font file and 2 pictures) but can contain any project resources we need.