trying to work on screen transitions
HomeScreen changed to OO, trying to get the transition to work from homescreen to levelscreen and back but is closing out when you try to return from levelscreen. Still need to work on clearing the screen and transitioning between views. If you look around line 40 in homescreen you can see how I am trying to design it